requirements.txt
appnope==0.1.0
asgiref==3.3.1
backcall==0.2.0
coverage==5.3
decorator==4.4.2
Django==3.1.3
django-extensions==3.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
python-decouple==3.3
pytz==2020.4
sqlparse==0.4.1
tqdm==4.54.0
traitlets==5.0.5
wcwidth==0.2.5
